Pass through unknown autocomplete config options
This allows the library to be forward-compatible in a JavaScript
environment when new fields are supported in the back-end without having
to update the library. The library must still be updated to include the
new fields in the interfaces, but forward-compatibility in TypeScript
can be achieved by casting the value passed in until the library can be
updated.